Update README.md
**Part 1: Analysed and Explored the Climate Data**
In this section, I have used Python and SQLAlchemy to do a basic climate analysis and data exploration of the climate database. Specifically, I have used SQLAlchemy ORM queries, Pandas, and Matplotlib. Below are the steps:


Used the SQLAlchemy create_engine() function to connect to the SQLite database.

Used the SQLAlchemy automap_base() function to reflect the tables into classes, and then saved references to the classes named station and measurement.

Linked Python to the database by creating a SQLAlchemy session.
Closed the session.
Perform a precipitation analysis and then a station analysis by completing the steps in the following two subsections.

**Precipitation Analysis**
Found the most recent date in the dataset.

Used the above date, got the previous 12 months of precipitation data by querying the previous 12 months of data and selecetd only the "date" and "prcp" values.

Loaded the query results into a Pandas DataFrame, and set the index to the "date" column.

Sorted the DataFrame values by "date".

Plotted the results by using the DataFrame by plot method.
Summary statistics for the precipitation data has been printed by using Pandas.
**Station Analysis**
Designed a query to calculate the total number of stations in the dataset.

Designed a query to find the most-active stations (that is, the stations that have the most rows) by completing the following steps:

Listed the stations and observation counts in descending order.

Answered the following question: which station id has the greatest number of observations?

Used the most-active station id, calculate the lowest, highest, and average temperatures.

Designed a query to get the previous 12 months of temperature observation (TOBS) data by completing the following steps:

Filtered by the station that has the greatest number of observations.

Queried the previous 12 months of TOBS data for that station.

Plotted the results as a histogram with bins=12.
Closed the session.
**Part 2: Design Your Climate App**
Designed a Flask API based on the queries that's beendeveloped. Used Flask to create the routes as follows:

/
Start at the homepage.

Listed all the available routes.

/api/v1.0/precipitation
Converted the query results to a dictionary by using date as the key and prcp as the value.

Returned the JSON representation of the dictionary.

/api/v1.0/stations
Returned a JSON list of stations from the dataset.
/api/v1.0/tobs
Queried the dates and temperature observations of the most-active station for the previous year of data.

Returned a JSON list of temperature observations for the previous year.

/api/v1.0/<start> and /api/v1.0/<start>/<end>
Returned a JSON list of the minimum temperature, the average temperature, and the maximum temperature for a specified start or start-end range.

For a specified start, calculated TMIN, TAVG, and TMAX for all the dates greater than or equal to the start date.

For a specified start date and end date, calculated TMIN, TAVG, and TMAX for the dates from the start date to the end date, inclusive.